Fix username case mismatch
If the user typed in their username in a different case than used
during registration, we would fail to persist their user object.  In
this rare case, we would continue to persist the previously logged-in
user.  Solve this by using a case-insensative comparison and always
clearing out the persisted user object when logging out.