Always persist client code
We need to be able to tell if an authentication token came from a
top-level account or a client code.  This isn't possible unless we
always persist the client code.

To test any chagnes here, use an app with the flag
moblico_login_show_client_code_accounts enabled, select an account,
and then exit the app at the login screen.  On the next app launch,
we won't correctly load the available clients unless we detected
that the client code was still set and reset the authentication token.